+++ src/main/java/noodlezip/badge/listener/CommentPostListener.java
@@ -1,10 +1,10 @@
-package noodlezip.badge.service.listeners;
+package noodlezip.badge.listener;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import noodlezip.badge.constants.LevelBadgeCategoryType;
 import noodlezip.badge.constants.UserEventType;
-import noodlezip.badge.events.BasicBadgeEvent;
+import noodlezip.badge.event.BasicBadgeEvent;
 import noodlezip.badge.service.process.level.LevelDirectUpdateProcessor;
 import org.springframework.scheduling.annotation.Async;
 import org.springframework.stereotype.Service;
@@ -13,25 +13,25 @@ import org.springframework.transaction.event.TransactionalEventListener;
 @Slf4j
 @RequiredArgsConstructor
 @Service
-public class CommunityLikeListener {
+public class CommentPostListener {
 
     private final LevelDirectUpdateProcessor directUpdateProcessor;
 
 
     @Async
     @TransactionalEventListener
-    public void processCommunityLikeCount(BasicBadgeEvent event) {
-        if (event.getEventType() == UserEventType.COMMUNITY_LIKE) {
+    public void processAllCommentPostCount(BasicBadgeEvent event) {
+        if (event.getEventType() == UserEventType.COMMENT_POST) {
 
             try {
                 directUpdateProcessor.process(
-                        event.getUserId(), LevelBadgeCategoryType.COMMUNITY_GET_LIKE_COUNT_BADGE);
+                        event.getUserId(), LevelBadgeCategoryType.ALL_COMMENT_POST_COUNT_BADGE);
 
             } catch (Exception e) {
                 log.error("[BadgeFail] userId={} event={} badgeType={} reason={}",
                         event.getUserId(),
-                        UserEventType.COMMUNITY_LIKE.name(),
-                        LevelBadgeCategoryType.COMMUNITY_GET_LIKE_COUNT_BADGE.name(),
+                        UserEventType.COMMENT_POST.name(),
+                        LevelBadgeCategoryType.ALL_COMMENT_POST_COUNT_BADGE.name(),
                         e.getMessage(), e
                 );
             }
