+++ src/main/java/noodlezip/badge/service/process/handler/ValueUpdateHandler.java
@@ -1,4 +1,4 @@
-package noodlezip.badge.service.process.level.handler;
+package noodlezip.badge.service.process.handler;
 
 import noodlezip.badge.constants.BadgeStrategyType;
 import noodlezip.badge.entity.UserBadge;
