+++ src/main/java/noodlezip/badge/service/process/level/RamenReviewSidoRegionBadge.java
@@ -10,8 +10,8 @@ import noodlezip.badge.entity.UserBadge;
 import noodlezip.badge.repository.BadgeCategoryRepository;
 import noodlezip.badge.repository.BadgeRepository;
 import noodlezip.badge.repository.UserBadgeRepository;
-import noodlezip.badge.service.process.level.handler.LevelUpHandler;
-import noodlezip.badge.service.process.level.handler.ValueUpdateHandler;
+import noodlezip.badge.service.process.handler.LevelUpHandler;
+import noodlezip.badge.service.process.handler.ValueUpdateHandler;
 import noodlezip.badge.status.BadgeErrorStatus;
 import noodlezip.common.exception.CustomException;
 import org.springframework.stereotype.Service;
